Add method to CounterReport to insert missing metrics
SUSHI protocol does not guarantee fields with zero use
are in a report. If a field has no use for a given
period, it might not show up in the report at all.

If a report is a datbase report, check each database
to ensure it has all metrics and insert them with a 0
for the first month of the reporting period. After all
metrics have been established, sort them so they write
to file in the correct order (sort order will be
maintained after the subsequent title sort)