Added tests and data files for DB reports through SUSHI
TestConvertRawDatabase tests that all information is
correct and in the correct order for a DB report.
Utilizes sushi_simple_db1.xml which is in part based
on formatting from ProQuest SUSHI reports

TestRawDatabaseWithMissingData tests that, on output,
SUSHI data is correct and full even if zero-value fields
were not sent. January is missing data and there are no
records for 'record_view' in the data source. Utilizes
sushi_db1_missing_record_view.xml which is in part based
on formatting from EBSCO SUSHI reports.